fix: 修复 Provider ID 类型溢出导致配置解析失败
**问题描述**
- Provider ID 字段使用 int 类型（32 位）
- 前端使用时间戳生成 ID（如 1763049954488）
- 大于 2^31-1 (2147483647) 的 ID 导致溢出
- 配置文件无法正确解析，所有 provider 被过滤
- 报错：没有可用的 provider 支持模型

**修复方案**
- 将 Provider.ID 类型从 int 改为 int64
- 更新相关的 map[int]string 为 map[int64]string
- 支持最大 ID 值：9223372036854775807

**影响范围**
- services/providerservice.go (2 处修改)
  - Line 13: ID 字段类型
  - Line 88: nameByID map 类型

**测试**
使用实际配置文件测试（ID: 1763049954488）
- ✅ 配置正确加载
- ✅ 模型支持检查正常
- ✅ 向后兼容老版本

**版本更新**
- 版本号：v0.3.2 → v0.3.3

+++ services/providerservice.go
@@ -10,7 +10,7 @@ import (
 )
 
 type Provider struct {
-	ID      int    `json:"id"`
+	ID      int64  `json:"id"` // 修复：使用 int64 支持大 ID 值
 	Name    string `json:"name"`
 	APIURL  string `json:"apiUrl"`
 	APIKey  string `json:"apiKey"`
@@ -85,7 +85,7 @@ func (ps *ProviderService) SaveProviders(kind string, providers []Provider) erro
 	if err != nil {
 		return err
 	}
-	nameByID := make(map[int]string, len(existingProviders))
+	nameByID := make(map[int64]string, len(existingProviders))
 	for _, p := range existingProviders {
 		nameByID[p.ID] = p.Name
 	}
